Renamed `JsonAbiNew` to `JsonAbi` and the old format to `JsonAbiOld`
