+++ packages/abi-coder/src/utils/transpile-abi.ts
@@ -4,8 +4,8 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 
-import type { JsonAbi } from '../types/JsonAbi';
-import type { JsonAbiNew } from '../types/JsonAbiNew';
+import type { JsonAbiOld } from '../types/JsonAbi';
+import type { JsonAbi } from '../types/JsonAbiNew';
 
 /**
  * This will transpile new ABIs to the old format.
@@ -18,7 +18,7 @@ import type { JsonAbiNew } from '../types/JsonAbiNew';
  * The old format contains only:
  *    - `types`
  */
-export function transpileAbi(abi: JsonAbiNew): JsonAbi {
+export function transpileAbi(abi: JsonAbi): JsonAbiOld {
   // do not transpile older versions
   if (!abi.specVersion) {
     return abi;
